wersje RZiS mozna wybierac

+++ trunk/npkpir_23/src/java/viewfk/PozycjaRZiSView.java
@@ -5,6 +5,7 @@
 package viewfk;
 
 import comparator.Kontocomparator;
+import converter.RomNumb;
 import daoFK.KontoDAOfk;
 import daoFK.KontoZapisyFKDAO;
 import daoFK.PozycjaRZiSDAO;
@@ -12,6 +13,7 @@ import embeddablefk.TreeNodeExtended;
 import entityfk.Konto;
 import entityfk.Kontozapisy;
 import entityfk.PozycjaRZiS;
+import entityfk.Rzisuklad;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -39,6 +41,7 @@ public class PozycjaRZiSView implements Serializable {
     private PozycjaRZiS selected;
     private ArrayList<TreeNodeExtended> finallNodes;
     private static ArrayList<PozycjaRZiS> pozycje;
+    private static ArrayList<PozycjaRZiS> pozycje_old;
     private static ArrayList<Konto> przyporzadkowanekonta;
     private List<Konto> wykazkont;
     @Inject
@@ -46,6 +49,7 @@ public class PozycjaRZiSView implements Serializable {
     private static String wybranapozycja;
     @Inject private KontoZapisyFKDAO kontoZapisyFKDAO;
     @Inject private PozycjaRZiSDAO pozycjaRZiSDAO;
+    @Inject private Rzisuklad rzisuklad;
     
 
     public PozycjaRZiSView() {
@@ -56,6 +60,7 @@ public class PozycjaRZiSView implements Serializable {
         this.przyporzadkowanekonta = new ArrayList<>();
         this.finallNodes = new ArrayList<TreeNodeExtended>();
         pozycje = new ArrayList<>();
+        pozycje_old = new ArrayList<>();
     }
 
     @PostConstruct
@@ -64,30 +69,30 @@ public class PozycjaRZiSView implements Serializable {
         zmodyfikujwykazkont(pobranekonta);
         Collections.sort(wykazkont, new Kontocomparator());
         //(int lp, String pozycjaString, String pozycjaSymbol, int macierzysty, int level, String nazwa, boolean przychod0koszt1, double kwota)
-//        pozycje.add(new PozycjaRZiS(1, "A", "A", 0, 0, "Przychody netto ze sprzedaży i zrównane z nimi, w tym:", false));
-//        pozycje.add(new PozycjaRZiS(2, "A.I", "I", 1, 1, "Przychody netto ze sprzedaży produktów", false, 0.0));
-//        pozycje.add(new PozycjaRZiS(3, "A.II", "II", 1, 1, "Zmiana stanu produktów", false, 0.0));
-//        pozycje.add(new PozycjaRZiS(4, "A.III", "III", 1, 1, "Koszt wytworzenia produktów na własne potrzeby jednostki", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(5, "A.IV", "IV", 1, 1, "Przychody netto ze sprzedaży towarów i materiałów", false, 0.0));
-//        pozycje.add(new PozycjaRZiS(6, "B", "B", 0, 0, "Koszty działalności operacyjnej", true));
-//        pozycje.add(new PozycjaRZiS(7, "B.I", "I", 6, 1, "Amortyzacja", true));
-//        pozycje.add(new PozycjaRZiS(8, "B.II", "II", 6, 1, "Zużycie materiałów i energii", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(9, "B.III", "III", 6, 1, "Usługi obce", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(10, "B.IV", "IV", 6, 1, "Podatki i  opłaty", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(11, "B.V", "V", 6, 1, "Wynagrodzenia", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(12, "B.I.1", "1", 7, 2, "amortyzacja kup", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(13, "B.I.2", "2", 7, 2, "amortyzacja nkup", true));
-//        pozycje.add(new PozycjaRZiS(14, "B.I.2.a)", "a)", 13, 3, "bobopo", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(15, "C", "C", 0, 0, "Zysk (strata) ze sprzedaży (A-B)", false, "A-B"));
-//        pozycje.add(new PozycjaRZiS(16, "D", "D", 0, 0, "Pozostałe truey operacyjne", false));
-//        pozycje.add(new PozycjaRZiS(17, "D.I", "I", 16, 1, "Zysk z niefinansowych aktywów trwałych", false, 0.0));
-//        pozycje.add(new PozycjaRZiS(18, "D.II", "II", 16, 1, "Dotacje", false, 0.0));
-//        pozycje.add(new PozycjaRZiS(19, "D.III", "III", 16, 1, "Inne truey operacyjne", false, 0.0));
-//        pozycje.add(new PozycjaRZiS(20, "E", "E", 0, 0, "Pozostałe koszty operacyjne", true));
-//        pozycje.add(new PozycjaRZiS(21, "E.I", "I", 20, 1, "Strata z niefinansowych aktywów trwałych", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(22, "E.II", "II", 20, 1, "Aktualizacja aktywów niefinansowych", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(23, "E.III", "III", 20, 1, "Inne koszty operacyjne", true, 0.0));
-//        pozycje.add(new PozycjaRZiS(24, "F", "F", 0, 0, "Zysk (strata) ze działalności operacyjnej (C+D-E)", false, "C+D-E"));
+        pozycje_old.add(new PozycjaRZiS(1, "A", "A", 0, 0, "Przychody netto ze sprzedaży i zrównane z nimi, w tym:", false));
+        pozycje_old.add(new PozycjaRZiS(2, "A.I", "I", 1, 1, "Przychody netto ze sprzedaży produktów", false, 0.0));
+        pozycje_old.add(new PozycjaRZiS(3, "A.II", "II", 1, 1, "Zmiana stanu produktów", false, 0.0));
+        pozycje_old.add(new PozycjaRZiS(4, "A.III", "III", 1, 1, "Koszt wytworzenia produktów na własne potrzeby jednostki", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(5, "A.IV", "IV", 1, 1, "Przychody netto ze sprzedaży towarów i materiałów", false, 0.0));
+        pozycje_old.add(new PozycjaRZiS(6, "B", "B", 0, 0, "Koszty działalności operacyjnej", true));
+        pozycje_old.add(new PozycjaRZiS(7, "B.I", "I", 6, 1, "Amortyzacja", true));
+        pozycje_old.add(new PozycjaRZiS(8, "B.II", "II", 6, 1, "Zużycie materiałów i energii", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(9, "B.III", "III", 6, 1, "Usługi obce", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(10, "B.IV", "IV", 6, 1, "Podatki i  opłaty", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(11, "B.V", "V", 6, 1, "Wynagrodzenia", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(12, "B.I.1", "1", 7, 2, "amortyzacja kup", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(13, "B.I.2", "2", 7, 2, "amortyzacja nkup", true));
+        pozycje_old.add(new PozycjaRZiS(14, "B.I.2.a)", "a)", 13, 3, "bobopo", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(15, "C", "C", 0, 0, "Zysk (strata) ze sprzedaży (A-B)", false, "A-B"));
+        pozycje_old.add(new PozycjaRZiS(16, "D", "D", 0, 0, "Pozostałe truey operacyjne", false));
+        pozycje_old.add(new PozycjaRZiS(17, "D.I", "I", 16, 1, "Zysk z niefinansowych aktywów trwałych", false, 0.0));
+        pozycje_old.add(new PozycjaRZiS(18, "D.II", "II", 16, 1, "Dotacje", false, 0.0));
+        pozycje_old.add(new PozycjaRZiS(19, "D.III", "III", 16, 1, "Inne truey operacyjne", false, 0.0));
+        pozycje_old.add(new PozycjaRZiS(20, "E", "E", 0, 0, "Pozostałe koszty operacyjne", true));
+        pozycje_old.add(new PozycjaRZiS(21, "E.I", "I", 20, 1, "Strata z niefinansowych aktywów trwałych", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(22, "E.II", "II", 20, 1, "Aktualizacja aktywów niefinansowych", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(23, "E.III", "III", 20, 1, "Inne koszty operacyjne", true, 0.0));
+        pozycje_old.add(new PozycjaRZiS(24, "F", "F", 0, 0, "Zysk (strata) ze działalności operacyjnej (C+D-E)", false, "C+D-E"));
         //tutaj dzieje sie magia :) tak funkcja przeksztalca baze danych w nody
         pozycje.addAll(pozycjaRZiSDAO.findAll());
         if (pozycje.size() == 0) {
@@ -99,8 +104,25 @@ public class PozycjaRZiSView implements Serializable {
             List<Konto> plankont = kontoDAO.findAll();
             ustawRoota(rootProjekt, pozycje, zapisy, plankont);
         }
+        List<Kontozapisy> zapisy = kontoZapisyFKDAO.findAll();
+        List<Konto> plankont = kontoDAO.findAll();
+        ustawRoota(root, pozycje_old, zapisy, plankont);
     }
 
+    public void pobierzuklad () {
+        pozycje = new ArrayList<>();
+        try {
+            pozycje.addAll(pozycjaRZiSDAO.findRzisuklad(rzisuklad));
+        } catch (Exception e){}
+        if (pozycje.size() == 0) {
+            pozycje.add(new PozycjaRZiS(1, "A", "A", 0, 0, "Kliknij tutaj i dodaj pierwszą pozycję RZiS", false));
+            Msg.msg("i", "Dodaje pusta pozycje");
+        }
+        rootProjekt = new TreeNodeExtended("root", null);
+        ustawRootaprojekt(rootProjekt, pozycje);
+        Msg.msg("i", "Pobrano układ "+rzisuklad.getRzisukladPK().getUklad());
+    }
+    
     private void ustawRoota(TreeNodeExtended rt, ArrayList<PozycjaRZiS> pz, List<Kontozapisy> zapisy, List<Konto> plankont) {
         rt.createTreeNodesForElement(pz);
         rt.addNumbers(zapisy, plankont);
@@ -275,7 +297,6 @@ public class PozycjaRZiSView implements Serializable {
     }
     
     public void dodajnowapozycje(String syntetycznaanalityczna) {
-        Msg.msg("i", "Dodaję nową pozycję w RZiS");
         if (syntetycznaanalityczna.equals("syntetyczna")){
             //dodaje nowa syntetyke
             if (pozycje.get(0).getNazwa().equals("Kliknij tutaj i dodaj pierwszą pozycję RZiS")) {
@@ -288,33 +309,101 @@ public class PozycjaRZiSView implements Serializable {
                 nowyelementRZiS.setLevel(0);
                 nowyelementRZiS.setMacierzysty(0);
             } else {
-                String poprzednialitera = pozycje.get(pozycje.size()-1).getPozycjaSymbol();
-                int charValue = poprzednialitera.charAt(0);
-                String nowalitera = String.valueOf( (char) (charValue + 1));
+                String poprzednialitera = ((PozycjaRZiS) rootProjekt.getChildren().get(rootProjekt.getChildCount()-1).getData()).getPozycjaSymbol();
+                String nowalitera = RomNumb.alfaInc(poprzednialitera);
                 nowyelementRZiS.setPozycjaSymbol(nowalitera);
                 nowyelementRZiS.setPozycjaString(nowalitera);
                 nowyelementRZiS.setLevel(0);
                 nowyelementRZiS.setMacierzysty(0);
+                if (!(nowyelementRZiS.getFormula() instanceof String)) {
+                    nowyelementRZiS.setFormula("");
+                }
                 Msg.msg("i", nowyelementRZiS.getNazwa()+"zachowam pod "+nowalitera);
             }
+            nowyelementRZiS.setUklad(rzisuklad.getRzisukladPK().getUklad());
+            nowyelementRZiS.setPodatnik(rzisuklad.getRzisukladPK().getPodatnik());
+            nowyelementRZiS.setRok(rzisuklad.getRzisukladPK().getRok());
             try {
                 pozycjaRZiSDAO.dodaj(nowyelementRZiS);
                 pozycje.add(nowyelementRZiS);
                 rootProjekt = new TreeNodeExtended("root", null);
                 ustawRootaprojekt(rootProjekt, pozycje);
-                Msg.msg("i", "Dodano nowa pozycję");
+                Msg.msg("i", "Dodano nowa pozycję syntetyczną");
             } catch (Exception e) {
-                Msg.msg("e", "Wystąpił błąd - nie dodano nowej pozycji");
+                Msg.msg("e", "Wystąpił błąd - nie dodano nowej pozycji syntetycznej");
             }
             nowyelementRZiS = new PozycjaRZiS();
             
         } else {
-            
+            if (pozycje.get(0).getNazwa().equals("Kliknij tutaj i dodaj pierwszą pozycję RZiS")) {
+                Msg.msg("e", "Błąd. Najpierw dodaj pierwszą pozycje wyższego rzędu!");
+                return;
+            }
+            int level = ((PozycjaRZiS) wybranynodekonta.getData()).getLevel();
+                if (level == 4) {
+                    Msg.msg("e", "Nie można dodawać więcej poziomów");
+                    return;
+                }
+            PozycjaRZiS parent = (PozycjaRZiS) wybranynodekonta.getData();
+            String nastepnysymbol;
+            //sprawdzic trzeba czy sa dzieci juz jakies
+            if (wybranynodekonta.getChildCount()==0) {
+                //w zaleznosci od levelu zwraca nastepny numer
+                nastepnysymbol = zwrocNastepnySymbol(level+1);
+            } else {
+                int index = wybranynodekonta.getChildCount()-1;
+                PozycjaRZiS lastchild = (PozycjaRZiS) wybranynodekonta.getChildren().get(index).getData();
+                nastepnysymbol = zwrocNastepnySymbol(level+1, lastchild.getPozycjaSymbol());
+            }
+            nowyelementRZiS.setPozycjaSymbol(nastepnysymbol);
+            nowyelementRZiS.setPozycjaString(parent.getPozycjaString()+"."+nastepnysymbol);
+            nowyelementRZiS.setLevel(level+1);
+            nowyelementRZiS.setMacierzysty(parent.getLp());
+            if (!(nowyelementRZiS.getFormula() instanceof String)) {
+                nowyelementRZiS.setFormula("");
+            }
+            nowyelementRZiS.setUklad(rzisuklad.getRzisukladPK().getUklad());
+            nowyelementRZiS.setPodatnik(rzisuklad.getRzisukladPK().getPodatnik());
+            nowyelementRZiS.setRok(rzisuklad.getRzisukladPK().getRok());
+            try {
+                pozycjaRZiSDAO.dodaj(nowyelementRZiS);
+                pozycje.add(nowyelementRZiS);
+                rootProjekt = new TreeNodeExtended("root", null);
+                ustawRootaprojekt(rootProjekt, pozycje);
+                Msg.msg("i", "Dodano nowa pozycję analityczną");
+            } catch (Exception e) {
+                Msg.msg("e", "Wystąpił błąd - nie dodano nowej pozycji analitycznej");
+            }
+            nowyelementRZiS = new PozycjaRZiS();
         }
+    }
+    
+    public String zwrocNastepnySymbol(int level) {
+        switch (level) {
+            case 1 : return "I";
+            case 2 : return "1";
+            case 3 : return "a";
+            case 4 : return "-";
+        }
+        return null;
+    }
+    
+     public String zwrocNastepnySymbol(int level, String pozycjasymbol) {
+        switch (level) {
+            case 1 : return RomNumb.romInc(pozycjasymbol);
+            case 2 : return RomNumb.numbInc(pozycjasymbol);
+            case 3 : return RomNumb.alfaInc(pozycjasymbol);
+            case 4 : return "-";
+        }
+        return null;
     }
     
     public void usunpozycje() {
         try {
+            if (wybranynodekonta.getChildCount() > 0) {
+                Msg.msg("w", "Wybrana pozycja RZiS zawiera podpunkty!");
+                throw new Exception();
+            }
             pozycje.remove(wybranynodekonta.getData());
             pozycjaRZiSDAO.destroy(wybranynodekonta.getData());
             if (pozycje.size() == 0) {
@@ -325,7 +414,7 @@ public class PozycjaRZiSView implements Serializable {
             ustawRootaprojekt(rootProjekt, pozycje);
             Msg.msg("i", "Usuwam w RZiS");
         } catch (Exception e) {
-             Msg.msg("i", "Nie udało się usunąć pozycji w RZiS");
+             Msg.msg("e", "Nie udało się usunąć pozycji w RZiS");
         }
     }
     
@@ -400,6 +489,15 @@ public class PozycjaRZiSView implements Serializable {
     public void setNowyelementRZiS(PozycjaRZiS nowyelementRZiS) {
         this.nowyelementRZiS = nowyelementRZiS;
     }
+
+    public Rzisuklad getRzisuklad() {
+        return rzisuklad;
+    }
+
+    public void setRzisuklad(Rzisuklad rzisuklad) {
+        this.rzisuklad = rzisuklad;
+    }
   
+    
     
 }
